server/requirements.txt
flask
flask-cors
pymongo
pandas
python-dotenv
openpyxl
requests

